chore: moves connector rhino one folder deeper, to match other connectors
