+++ ConnectorRhino/.editorconfig
@@ -1,0 +1,79 @@
+root = true
+# EditorConfig is awesome:http://EditorConfig.org
+
+# top-most EditorConfig file
+
+# Don't use tabs for indentation.
+[*]
+indent_style = space
+indent_size = 2
+
+# Code files
+[*.{cs, csx, vb, vbx}]
+insert_final_newline = true
+charset = utf-8-bom
+
+# ReSharper properties
+resharper_csharp_extra_spaces = leave_multiple
+resharper_csharp_naming_rule.private_constants = AaBb, AA_BB
+resharper_csharp_space_within_array_access_brackets = true
+resharper_space_around_arrow_op = true
+resharper_space_within_array_rank_brackets = true
+resharper_use_indent_from_vs = false
+resharper_xml_attribute_style = first_attribute_on_single_line
+resharper_xml_pi_attribute_style = first_attribute_on_single_line
+
+# Dotnet code style settings:
+[*.{cs, vb}]
+# Sort using and Import directives with System.* appearing first
+dotnet_sort_system_directives_first = true
+# Avoid "this." and "Me." if not necessary
+dotnet_style_qualification_for_field = false : suggestion
+dotnet_style_qualification_for_property = false : suggestion
+dotnet_style_qualification_for_method = false : suggestion
+dotnet_style_qualification_for_event = false : suggestion
+
+# Use language keywords instead of framework type names for type references
+dotnet_style_predefined_type_for_locals_parameters_members = true : suggestion
+dotnet_style_predefined_type_for_member_access = true : suggestion
+
+# Suggest more modern language features when available
+dotnet_style_object_initializer = true : suggestion
+dotnet_style_collection_initializer = true : suggestion
+dotnet_style_coalesce_expression = true : suggestion
+dotnet_style_null_propagation = true : suggestion
+dotnet_style_explicit_tuple_names = true : suggestion
+
+# CSharp code style settings:
+[*.cs]
+# Prefer "var" everywhere
+csharp_style_var_for_built_in_types = true : suggestion
+csharp_style_var_when_type_is_apparent = true : suggestion
+csharp_style_var_elsewhere = true : suggestion
+
+# Prefer method-like constructs to have a block body
+csharp_style_expression_bodied_methods = false : none
+csharp_style_expression_bodied_constructors = false : none
+csharp_style_expression_bodied_operators = false : none
+
+# Prefer property-like constructs to have an expression-body
+csharp_style_expression_bodied_properties = true : none
+csharp_style_expression_bodied_indexers = true : none
+csharp_style_expression_bodied_accessors = true : none
+
+# Suggest more modern language features when available
+csharp_style_pattern_matching_over_is_with_cast_check = true : suggestion
+csharp_style_pattern_matching_over_as_with_null_check = true : suggestion
+csharp_style_inlined_variable_declaration = true : suggestion
+csharp_style_throw_expression = true : suggestion
+csharp_style_conditional_delegate_call = true : suggestion
+
+# Newline settings
+csharp_new_line_before_open_brace = all
+csharp_new_line_before_else = true
+csharp_new_line_before_catch = true
+csharp_new_line_before_finally = true
+csharp_new_line_before_members_in_object_initializers = true
+csharp_new_line_before_members_in_anonymous_types = true
+
+spaces_around_brackets = inside
